Add update support.
Build InstallBuilder autoupdater binaries, install as an optional
component including an update.ini to control it, and add a Start Menu
entry to access it. Also generate an update.xml file to go in
sirikata.com/releases, which is polled to check for updates.

We also need to generate an extra version of the mac installer because
the autoupdater needs a .tgz instead of the .dmg we were already
producing.